pyproject.toml
[tool.poetry]
name = "autodecrypt"
version = "2.0.9"
description = "Tool to decrypt iOS firmware images"
authors = ["matteyeux"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/matteyeux/autodecrypt"
repository = "https://github.com/matteyeux/autodecrypt"
keywords = ['autodecrypt', 'iOS', 'iBoot', 'checkm8']
classifiers=[
	'Development Status :: 5 - Production/Stable',
	'Intended Audience :: Developers',
	'Intended Audience :: Information Technology',
	'Topic :: Utilities',
	'License :: OSI Approved :: MIT License',
	'Programming Language :: Python :: 3.6',
	'Programming Language :: Python :: 3.7',
	'Programming Language :: Python :: 3.8',
]

[tool.poetry.dependencies]
python = "^3.7"
remotezip = "^0.9.2"
pyquery = "^1.4.1"
pre-commit = "^2.5.1"
pyusb = "^1.0.2"
beautifulsoup4 = "^4.9.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.5.1"

[tool.poetry.scripts]
autodecrypt = "autodecrypt.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
